+++ jt-808-server-spring-boot-autoconfigure/src/main/java/io/github/hylexus/jt/jt808/boot/props/Jt808ServerProps.java
@@ -4,7 +4,7 @@ import io.github.hylexus.jt.jt808.boot.props.builtin.BuiltComponentsProps;
 import io.github.hylexus.jt.jt808.boot.props.builtin.RequestSubPackageStorageProps;
 import io.github.hylexus.jt.jt808.boot.props.builtin.ResponseSubPackageStorageProps;
 import io.github.hylexus.jt.jt808.boot.props.msg.processor.MsgProcessorProps;
-import io.github.hylexus.jt.jt808.boot.props.plugin.Jt808PluginProps;
+import io.github.hylexus.jt.jt808.boot.props.feature.Jt808FeatureProps;
 import io.github.hylexus.jt.jt808.boot.props.protocol.ProtocolConfig;
 import io.github.hylexus.jt.jt808.boot.props.server.Jt808NettyTcpServerProps;
 import lombok.Getter;
@@ -55,5 +55,5 @@ public class Jt808ServerProps {
     private MsgProcessorProps msgProcessor = new MsgProcessorProps();
 
     @NestedConfigurationProperty
-    private Jt808PluginProps plugins = new Jt808PluginProps();
+    private Jt808FeatureProps features = new Jt808FeatureProps();
 }